requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-community
chromadb
sentence-transformers
ollama
pydantic
python-multipart
streamlit
requests
rank_bm25
langchain-openai
python-dotenv
